Add paging, but I need to load kernel at 0x100000 not at 0x1000
Don't load kernel at low memory bois :flushed:

+++ latest/lib/memlib.h
@@ -4,7 +4,7 @@
 #include "../lib/types.h"
 
 void memory_copy(char *source, char *dest, int nbytes);
-void memory_set(u8 *dest, u8 val, u32 len);
+void memory_set(u32 *dest, u8 val, u32 len);
 u32 kmalloc(u32 size, int align, u32 *phys_addr);
 
 #endif
